rename gear_shop_helper to lib_gear
Thank you to Kat for validating that the helper library is in fact used
elsewhere; that means it needs a better name. This commit renames it
as follows:

- overlay name: `gear_shop_helper` => `lib_gear`
- function name prefix: `GearShopHelper` => `LibGear`
- source directory: `src/gear_shop_helper` => `src/lib_gear`

additional changes:
- sort `config/symbol_addrs.slus_006.64.txt` by memory address
- remove size from `g_Menu` because `g_Menu` is a pointer, not the actual struct
- fix size of `_spu_dummy`
- fix size of `g_SoundVolumeController`
- update `SoundVolumeController` to remove the `cdInterpolator` field
- define `g_SoundCdAudioInterpolator` to capture the removed field
- update code to use `g_SoundCdAudioInterpolator` instead of `g_SoundVolumeController.cdInterpolator`
- various updates to migrate jump table rodata; does not eliminate the "rodata segment foo may belong
  to text segment bar" messages, but reduces them
- also address file split messages, resulting in splitting `system/temp1.c` to create `system/temp2.c`
