+++ config/checksum.sha
@@ -2,5 +2,5 @@ dc0b2dd786203d4cce5927c5a3fc85a18f39a3f7406078860076ebb0bbae7119  build/out/slus
 38a1ce829a6f094c505f67143d6ace2d328418c65425a7383991179467e1fdfc  build/out/field.bin
 3b9e2b890c27ae0de97fe343ac75cd35c05fe7f9605ae387d2166da0be78109c  build/out/member_change_menu.bin
 7890e14bcabddcf85368de10783ecff8daa166dc5ac254e06db5e27959cab4cf  build/out/shop_menu.bin
-14395a9f54c124fed016f07906cc882b2a9256d5ade2d10eef99e10fe9366523  build/out/gear_shop_helper.bin
+14395a9f54c124fed016f07906cc882b2a9256d5ade2d10eef99e10fe9366523  build/out/lib_gear.bin
 c7a1982f0ea624c81fd0f85e1a1335c7517723f12ba3e2d08395842392423691  build/out/gear_shop_menu.bin
