+++ config/field.yaml
@@ -54,10 +54,22 @@ segments:
     subalign: 4
     subsegments:
       - [0x0, rodata]
+      - [0x4, .rodata, main/misc3]
+      - [0x14, .rodata, main/misc2]
+      - [0x90, .rodata, main/main]
+      - [0x9C, .rodata, main/misc4]
+      - [0x11C, .rodata, main/misc8]
+      - [0x198, .rodata, main/misc]
       - [0x1B8, .rodata, main/misc11]
-      - [0x1D0, rodata]
+      - [0x1D0, .rodata, main/misc11]
+      - [0x22C, .rodata, main/misc7]
+      - [0x240, .rodata, main/misc12]
+      - [0x254, rodata]
       - [0x268, .rodata, scripts/virtual_machine]
       - [0x294, rodata]
+      - [0x2BC, .rodata, main/misc5]
+      - [0x2D8, rodata]
+      - [0x2E8, .rodata, main/misc9]
       - [0x2FC, c, main/misc3]
       - [0x1F74, c, effects/fade]
       - [0x23F8, c, main/init]
@@ -76,6 +88,7 @@ segments:
       - [0x26B54, c, game_logic/scenario_flags]
       - [0x26D54, c, party/stats]
       - [0x27804, c, main/misc7]
+      - [0x29FD0, c, main/misc12]
       - [0x2C01C, c, dialogue/text_box]
       - [0x2D480, c, scripts/variable_handlers]
       - [0x2DF2C, c, main/misc6]
