+++ include/field/script_vm.h
@@ -26,7 +26,7 @@ typedef void (*ScriptVMHandler)(void);
 
 // Is Yielding maybe?
 extern int D_800B00C0;
-extern char D_8006FD84;
+extern char D_8006FD84[];
 extern int D_800ADBE0;
 extern int D_800ADBE4;
 extern int D_800ADBEC;
