+++ src/field/main/misc12.c
@@ -1,0 +1,284 @@
+#include "common.h"
+#include "main/game.h"
+#include "system/math.h"
+#include "field/main.h"
+#include "field/actor.h"
+#include "field/camera.h"
+#include "field/script_vm.h"
+#include "field/text_box.h"
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_80099AC0);
+
+void FieldScriptVMWriteCurCharacterID(void) {
+    FieldScriptMemoryWriteU16(SCRIPT_IMM_ARG(1), g_FieldScriptVMCurActor->characterId);
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 3;
+}
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", FieldScriptVMWritePartyLeaderCharacterID);
+
+void FieldScriptVMHandlerGetActorDirection(void) {
+    int delta = PSX_DEGREES(22.5);
+
+    // The direction is likely offset by two due to how g_FieldAngleToDirectionLUT is set up.
+    FieldScriptMemoryWriteU16(
+        SCRIPT_IMM_ARG(1),
+        (PSX_ANGLE_TO_DIRECTION_8(g_FieldScriptVMCurActor->rotation.vy + delta) + 2) & MASK_8DIR_MOVEMENT_NUM_DIRECTIONS
+    );
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 3;
+}
+
+void FieldScriptVMHandlerGetActorPosition(void) {
+    int nActorIndex = FieldScriptVMGetActorIndex(1);
+    if (nActorIndex != ACTOR_ID_INVALID) {
+        FieldScriptMemoryWriteU16(SCRIPT_IMM_ARG_ALIGNED(1), g_FieldActors[nActorIndex].childMatrix.t[0]);
+        FieldScriptMemoryWriteU16(SCRIPT_IMM_ARG_ALIGNED(2), g_FieldActors[nActorIndex].childMatrix.t[2]);
+        FieldScriptMemoryWriteU16(SCRIPT_IMM_ARG_ALIGNED(3), g_FieldActors[nActorIndex].childMatrix.t[1]);
+    }
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 8;
+}
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009A0FC);
+
+void FieldScriptVMHandlerPlayAnimation(void) {
+    unsigned char animationID;
+
+    g_FieldScriptVMCurActor->flags &= 0xFEFFFFFF;
+    animationID = *(u8*)&g_FieldScriptVMCurScriptData[g_FieldScriptVMCurActor->scriptInstructionPointer + 1];
+    g_FieldScriptVMCurActor->unkAnimationId = animationID;
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 2;
+}
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009A174);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009A1AC);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009A1E4);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009A2A8);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009A34C);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009A420);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009A490);
+
+int FieldGetCameraDirection(void) {
+    int halfDirection = PSX_DEGREES(22.5);
+    return (7 - PSX_ANGLE_TO_DIRECTION_8(g_CamInterpolation.targetAngleY - halfDirection))
+        & MASK_8DIR_MOVEMENT_NUM_DIRECTIONS;
+}
+
+void FieldScriptWriteCameraDirection(void) {
+    FieldScriptMemoryWriteU16(SCRIPT_IMM_ARG(1), FieldGetCameraDirection() & 0xFFFF);
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 3;
+}
+
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009A58C);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009A5E0);
+
+void FieldScriptSetDollySet(void) {
+    g_Scene.dollySet = FieldScriptVMGetArgument(1);
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 3;
+}
+
+void FieldScriptSetDollyStop(void) {
+    g_Scene.dollyStop = FieldScriptVMGetArgument(1);
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 3;
+}
+
+void FieldScriptCos(void) {
+    int nValue;
+    int angle;
+    int nFactor;
+    int address;
+
+    address = SCRIPT_IMM_ARG(1);
+    angle = FieldScriptArgument2(ARG(2), SCRIPT_READ_U8_REL(7));
+    nFactor = FieldScriptArgument3(ARG(3),  SCRIPT_READ_U8_REL(7)); // ?
+    nValue = rcos(angle) * nFactor;
+    FieldScriptMemoryWriteU16(address, nValue >> 12);
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 8;
+}
+
+void FieldScriptSin(void) {
+    int nValue;
+    int angle;
+    int nFactor;
+    int address;
+
+    address = SCRIPT_IMM_ARG(1);
+    angle = FieldScriptArgument2(ARG(2), SCRIPT_READ_U8_REL(7));
+    nFactor = FieldScriptArgument3(ARG(3),  SCRIPT_READ_U8_REL(7));
+    nValue = rsin(angle) * nFactor;
+    FieldScriptMemoryWriteU16(address, nValue >> 12);
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 8;
+}
+
+void FieldScriptAtan2(void) {
+    int x;
+    int y;
+    int address;
+
+    address = SCRIPT_IMM_ARG(1);
+    y = FieldScriptArgument2(ARG(2), SCRIPT_READ_U8_REL(7));
+    x = FieldScriptArgument3(ARG(3), SCRIPT_READ_U8_REL(7));
+
+    FieldScriptMemoryWriteU16(
+        address & 0xFFFF,
+        (s16) ratan2(y, x)
+    );
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 8;
+}
+
+int FieldScriptGetCurActorDirection(void) {
+    // The direction is likely offset by two due to how g_FieldAngleToDirectionLUT is set up.
+    int halfDirection = PSX_DEGREES(22.5);
+    return (PSX_ANGLE_TO_DIRECTION_8(g_FieldScriptVMCurActor->rotation.vy + halfDirection) + 2) & MASK_8DIR_MOVEMENT_NUM_DIRECTIONS;
+}
+
+extern s32 D_800ADB1C; // Is current actor a 2D actor (0) maybe?
+void FieldSetCurrentActorRotation(int rotation) {
+    short rotationValue2D;
+    short rotationValue3D;
+
+    if (D_800ADB1C == 0) {
+        rotationValue3D = rotation | 0x8000;
+        g_FieldScriptVMCurActor->rotation.vx = rotationValue3D;
+        g_FieldScriptVMCurActor->rotation.vy = rotationValue3D;
+        g_FieldScriptVMCurActor->rotation.vz = rotationValue3D;
+    }
+    rotationValue2D = rotation | 0x8000;
+    g_FieldScriptVMCurActor->rotation.vx = rotationValue2D;
+    g_FieldScriptVMCurActor->rotation.vy = rotationValue2D;
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 3;
+}
+
+void FieldScriptSetActorRotation(int angle) {
+    ActorData* pActor;
+    int actorIndex;
+    int rotAngle3D;
+    int rotAngle2D;
+
+    actorIndex = FieldScriptVMGetActorIndex(1);
+    if (actorIndex != ACTOR_ID_INVALID) {
+        pActor = g_FieldActors[FieldScriptVMGetActorIndex(1)].pActorData;
+        if (D_800ADB1C == 0) {
+            rotAngle3D = angle | 0x8000;
+            pActor->rotation.vx = rotAngle3D;
+            pActor->rotation.vy = rotAngle3D;
+            pActor->rotation.vz = rotAngle3D;
+
+        }
+        rotAngle2D = angle | 0x8000;
+        pActor->rotation.vx = rotAngle2D;
+        pActor->rotation.vy = rotAngle2D;
+    }
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 4;
+}
+
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009AA00);
+
+void func_8009AB08(int rotation) {
+    short rotationValue;
+
+    rotationValue = PSX_ANGLE(rotation - g_CamInterpolation.targetAngleY) | 0x8000;
+    g_FieldScriptVMCurActor->rotation.vx = rotationValue;
+    g_FieldScriptVMCurActor->rotation.vy = rotationValue;
+
+    // 3D Actor?
+    if (D_800ADB1C == 0) {
+        g_FieldScriptVMCurActor->rotation.vz = rotationValue;
+    }
+
+    g_FieldScriptVMCurActor->scriptInstructionPointer += 3;
+}
+
+
+extern u16 g_FieldAngleToDirectionLUT[
+    // PSX_DEGREES(270) | 0x8000,
+    // PSX_DEGREES(315) | 0x8000,
+    // PSX_DEGREES(0) | 0x8000,
+    // PSX_DEGREES(45) | 0x8000,
+    // PSX_DEGREES(90) | 0x8000,
+    // PSX_DEGREES(135) | 0x8000,
+    // PSX_DEGREES(180) | 0x8000,
+    // PSX_DEGREES(225) | 0x8000,
+];
+
+// Rotate current actor N turns clockwise, where 1 turn is 45 degrees
+void FieldScriptRotateActorClockwise(void) {
+    int numTurns = FieldScriptVMGetArgument(1);
+    FieldSetCurrentActorRotation(g_FieldAngleToDirectionLUT[
+        numTurns + FieldScriptGetCurActorDirection() & MASK_8DIR_MOVEMENT_NUM_DIRECTIONS
+    ]);
+}
+
+// Rotate current actor N turns counter-clockwise, where 1 turn is 45 degrees
+void FieldScriptRotateActorCounterClockwise(void) {
+    int numTurns = FieldScriptVMGetArgument(1);
+    FieldSetCurrentActorRotation(g_FieldAngleToDirectionLUT[
+        FieldScriptGetCurActorDirection() - numTurns & MASK_8DIR_MOVEMENT_NUM_DIRECTIONS
+    ]);
+}
+
+void FieldScriptSetActorDirection(void) {
+    int directionIndex = FieldScriptVMGetArgument(2);
+    FieldScriptSetActorRotation(g_FieldAngleToDirectionLUT[directionIndex]);
+}
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009AC34);
+
+void FieldScriptVMHandlerSetCurActorRotation(void) {
+    int directionIndex = FieldScriptVMGetArgument(1);
+    FieldSetCurrentActorRotation(g_FieldAngleToDirectionLUT[directionIndex]);
+}
+
+void func_8009ACB4(void) {
+    int directionIndex = FieldScriptVMGetArgument(1);
+    func_8009AB08(g_FieldAngleToDirectionLUT[directionIndex]);
+}
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009ACEC);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009AD6C);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009ADDC);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009AE0C);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009AE3C);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009AEE0);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B15C);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B184);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B210);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B338);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B398);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B664);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B6AC);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B708);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B7A8);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B824);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B884);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B8E4);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009B9A0);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009BA0C);
+
+INCLUDE_ASM("asm/field/nonmatchings/main/misc12", func_8009BA7C);
