+++ src/slus_006.64/system/rendering.c
@@ -95,7 +95,7 @@ INCLUDE_ASM("asm/slus_006.64/nonmatchings/system/rendering", func_8001F6B0);
 INCLUDE_ASM("asm/slus_006.64/nonmatchings/system/rendering", func_8001F750);
 
 // Animation frame and tile data stuff
-INCLUDE_ASM("asm/slus_006.64/nonmatchings/system/rendering", func_8001F8E8)
+INCLUDE_ASM("asm/slus_006.64/nonmatchings/system/rendering", func_8001F8E8);
